fix(restructure-controller): Added correct types
