+++ server/src/utils/authUtils.ts
@@ -1,7 +1,6 @@
 // File for providing authenticaion based util functions
-import { Permission } from "../global/globalTypes";
-import * as authTypes from "../types/authTypes";
-import * as globalTypes from "../global/globalTypes";
+
+import { JWTPayload, Permission, User } from "../types/authTypes";
 
 /**
  * Checks if a list of permissions of a member include a list of permission IDs
@@ -54,7 +53,7 @@ export const doesRolesHaveSomeOf = (memberRoles: number[], roles: number[]) => {
  * @param directorPermissionsResult The result of the director permissions query
  * @returns The payload for the user data
  */
-export const createUserDataPayload = (result: authTypes.User, directorPermissionsResult: Permission[]) => {
+export const createUserDataPayload = (result: User, directorPermissionsResult: Permission[]) => {
   const permissions = [];
   const roles = [];
   // Adds role permissions to the permissions array and adds directorID to the roles array
@@ -78,7 +77,7 @@ export const createUserDataPayload = (result: authTypes.User, directorPermission
         permissions.push({ permissionID: perm, canDelegate: 0 });
       });
   }
-  const payload: globalTypes.JWTPayload = {
+  const payload: JWTPayload = {
     mitgliedID: result.mitgliedID,
     name: result.name,
     permissions,
